requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
loguru
requests
pytest
httpx
pytest-asyncio
pytest-cov
playwright
pytest-playwright
openai
google-cloud-datastore
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
Pillow
yt-dlp
email-validator

google-auth-oauthlib
google-api-python-client
moviepy==1.0.3
mutagen

APScheduler==3.10.4
slowapi
stable-ts
